fix(core): split scanner.rs's HEADER-skip into a sibling module to clear the merge-inflated module-size budget
Merging upstream/main into this branch combined #3673's four-scanner
comment-trivia work with this PR's own unterminated-record reporting,
neither of which alone exceeded the 601-line budget recorded for
rust/core/src/parser/scanner.rs, but together landed it at 606 —
5 over, and `no_module_grows_past_its_ratchet_budget` failed on it.

`data_section_start` (the STEP HEADER-skip) is a self-contained free
function with no scanner state, so it moves verbatim to a new sibling
module, scanner_header.rs (62 lines), leaving scanner.rs at 557 lines
and the allowlist budget lowered to match rather than raised.
