+++ rust/core/src/parser/scanner.rs
@@ -7,6 +7,10 @@
 //! Independent of the nom [`tokenizer`](super::tokenizer): does its own
 //! hand-rolled, quote- and comment-aware parsing without building [`Token`]s.
 
+#[path = "scanner_header.rs"]
+mod scanner_header;
+use scanner_header::data_section_start;
+
 /// Fast entity scanner over raw IFC bytes without full parsing.
 /// O(n) performance for finding entities by type
 /// Uses memchr for SIMD-accelerated byte searching
@@ -546,59 +550,6 @@ where
     T: AsRef<[u8]> + ?Sized,
 {
     EntityScanner::new(content).count()
-}
-
-/// Locate the byte offset of the first character after `DATA;` (skipping the
-/// STEP HEADER section). Returns 0 if the marker isn't found — partial files
-/// without a HEADER still scan from the top.
-///
-/// Scanning the HEADER for entities is unsafe: the HEADER is a free-form
-/// STEP record that legally contains arbitrary characters inside quoted
-/// strings (filenames, descriptions). CATIA emits `FILE_NAME('…\X0\2#.ifc'…)`,
-/// and a tokenizer that anchors on `#` will latch onto the in-string `#`,
-/// flip `find_entity_end`'s quote parity, and drop the rest of the file.
-/// See issue #654.
-///
-/// Quote-aware: the marker is only matched outside `'…'` strings, since a
-/// HEADER field could legally contain the literal text `DATA;` in a
-/// description or filename. Escaped single quotes (`''`) are treated as a
-/// pair of in-string characters per ISO 10303-21.
-fn data_section_start(bytes: &[u8]) -> usize {
-    const MARKER: &[u8] = b"DATA;";
-    let len = bytes.len();
-    if len < MARKER.len() {
-        return 0;
-    }
-    // Cap the header scan. Real-world headers are <2 KB; an unbounded scan
-    // here would defeat the point of an O(1)-up-front fix on giant files
-    // that legitimately lack a HEADER section.
-    let limit = len.min(1 << 18); // 256 KB
-    let mut pos = 0;
-    let mut in_string = false;
-    while pos < limit {
-        let b = bytes[pos];
-        if in_string {
-            if b == b'\'' {
-                if pos + 1 < limit && bytes[pos + 1] == b'\'' {
-                    pos += 2; // escaped quote
-                    continue;
-                }
-                in_string = false;
-            }
-            pos += 1;
-            continue;
-        }
-        if b == b'\'' {
-            in_string = true;
-            pos += 1;
-            continue;
-        }
-        if b == b'D' && pos + MARKER.len() <= len && &bytes[pos..pos + MARKER.len()] == MARKER {
-            return pos + MARKER.len();
-        }
-        pos += 1;
-    }
-    0
 }
 
 #[cfg(test)]
